+++ web/main.dart
@@ -27,7 +27,7 @@ FrameCounter _counter;
  *
  * The current [time] is passed in.
  */
-void _onUpdate(int time)
+void _onUpdate(double time)
 {
   _counter.update(time);
   Game.onUpdate(time);
